Got conditional compiling working on Ubuntu 23
