+++ CMakeLists.txt
@@ -78,10 +78,13 @@ set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} ${JSON_CFLAGS}")
 set(JUDY_LIBRARIES "Judy")
 
 pkg_check_modules(LIBURING liburing)
+set(WITH_LIBURING 0)
 if(NOT APPLE AND NOT BSD AND NOT WIN32 AND NOT NetBSD AND LIBURING_FOUND)
-    set(WITH_LIBURING TRUE)
+    set(WITH_LIBURING 1)
     set(LIBURING_LIBRARIES uring)
 endif()
+add_definitions(-DWITH_LIBURING=${WITH_LIBURING})
+
 # Standard FLAGS
 set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -std=gnu99")
 if(NOT APPLE)
